+++ devhub/projects/udemy_react_the_complete_guide_2024/s03/src/App.jsx
@@ -1,35 +1,7 @@
-import reactImg from "./assets/react-core-concepts.png";
 import componetsImg from "./assets/components.png";
-import { CORE_CONCEPTS } from "./data";
-
-const reactDescription = ["Fundamental", "Crucial", "Core"];
-function genRandomInt(max) {
-  return Math.floor(Math.random() * (max + 1));
-}
-
-function Header() {
-  const description = reactDescription[genRandomInt(2)];
-  return (
-    <header>
-      <img src={reactImg} alt="Stylized atom" />
-      <h1>React Essentials</h1>
-      <p>
-        {description} React concepts you will need for almost any app you are
-        going to build!
-      </p>
-    </header>
-  );
-}
-
-function CoreConcept({ image, title, description }) {
-  return (
-    <li>
-      <img src={image} alt={title} />
-      <h3>{title}</h3>
-      <p>{description}</p>
-    </li>
-  );
-}
+import { CORE_CONCEPTS } from "./data.js";
+import Header from "./components/Header.jsx";
+import CoreConcept from "./components/CoreConcept.jsx";
 
 function App() {
   return (
