+++ devhub/projects/udemy_react_the_complete_guide_2024/s03/src/components/CoreConcept.jsx
@@ -1,0 +1,9 @@
+export default function CoreConcept({ image, title, description }) {
+  return (
+    <li>
+      <img src={image} alt={title} />
+      <h3>{title}</h3>
+      <p>{description}</p>
+    </li>
+  );
+}
