+++ devhub/projects/udemy_react_the_complete_guide_2024/s03/src/components/Header.jsx
@@ -1,0 +1,20 @@
+import reactImg from "../assets/react-core-concepts.png";
+
+const reactDescription = ["Fundamental", "Crucial", "Core"];
+function genRandomInt(max) {
+  return Math.floor(Math.random() * (max + 1));
+}
+
+export default function Header() {
+  const description = reactDescription[genRandomInt(2)];
+  return (
+    <header>
+      <img src={reactImg} alt="Stylized atom" />
+      <h1>React Essentials</h1>
+      <p>
+        {description} React concepts you will need for almost any app you are
+        going to build!
+      </p>
+    </header>
+  );
+}
